Fixed a range error in HostAddress test

+++ include/address.h
@@ -27,7 +27,7 @@ public:
   HostAddress(const sockaddr_storage& addr);
   HostAddress(const sockaddr_in& addr);
   HostAddress(const sockaddr_in6& addr);
-  HostAddress(const std::string& ip, const int& port);
+  HostAddress(const std::string& ip, const uint16_t& port);
   ~HostAddress();
 
   friend std::ostream& operator<<(std::ostream& out, const HostAddress& addr);
@@ -35,7 +35,7 @@ public:
   void setAddress(const sockaddr_storage& addr);
   void setAddress(const sockaddr_in& addr);
   void setAddress(const sockaddr_in6& addr);
-  void setAddress(const std::string& ip, const int& port);
+  void setAddress(const std::string& ip, const uint16_t& port);
   Version version() const;
   bool isValid() const;
   std::string ipString() const;
